+++ pumpkin-lib/src/basic_types/sequence_generators/sequence_generator.rs
@@ -1,0 +1,5 @@
+use std::fmt::Debug;
+
+pub trait SequenceGenerator: Debug {
+    fn next(&mut self) -> i64;
+}