+++ pumpkin-lib/src/engine/sat/mod.rs
@@ -1,16 +1,16 @@
 mod assignments_propositional;
 pub mod clause_allocators;
 mod explanation_clause_manager;
-mod glucose_restart_strategy;
 mod learned_clause_manager;
 mod learned_clause_minimiser;
+mod restart_strategy;
 mod sat_engine_data_structures;
 
 pub use assignments_propositional::AssignmentsPropositional;
 pub use explanation_clause_manager::ExplanationClauseManager;
-pub use glucose_restart_strategy::GlucoseRestartStrategy;
 pub use learned_clause_manager::LearnedClauseManager;
 pub use learned_clause_manager::SatOptions;
 pub use learned_clause_minimiser::LearnedClauseMinimiser;
+pub use restart_strategy::RestartStrategy;
 pub use sat_engine_data_structures::LearnedClauseSortingStrategy;
 pub use sat_engine_data_structures::SATEngineDataStructures;